sfeed_update: remove xargs -s
Theres no need to specify. POSIX defines it should support at least LINE_MAX
(2048 typically).

OpenIndiana xargs doesn't conform to POSIX. It doesn't use the largest constraint but
errors out.

From POSIX:
https://pubs.opengroup.org/onlinepubs/9699919799/utilities/xargs.html

"Values of size up to at least {LINE_MAX} bytes shall be supported, provided
that the constraints specified in the DESCRIPTION are met. It shall not be
considered an error if a value larger than that supported by the implementation
or exceeding the constraints specified in the DESCRIPTION is given; xargs shall
use the largest value it supports within the constraints."